+++ config/explorer.php
@@ -21,4 +21,10 @@ return [
     'indexes' => [
         // \App\Models\Post::class
     ],
+
+    /**
+     * You may opt to keep the old indices after the alias is pointed to a new index.
+     * A model is only using index aliases if it implements the Aliased interface.
+     */
+    'prune_old_aliases' => true,
 ];