+++ src/Domain/IndexManagement/IndexConfigurationInterface.php
@@ -8,9 +8,15 @@ interface IndexConfigurationInterface
 {
     public function getName(): string;
 
+    public function isAliased(): bool;
+
+    public function getAliasConfiguration(): IndexAliasConfigurationInterface;
+
     public function getProperties(): array;
 
     public function getSettings(): array;
 
     public function toArray(): array;
+
+    public function getConfiguredIndexName(): string;
 }