+++ src/ExplorerServiceProvider.php
@@ -46,7 +46,10 @@ class ExplorerServiceProvider extends ServiceProvider
         });
 
         $this->app->bind(IndexConfigurationRepositoryInterface::class, function () {
-            return new ElasticIndexConfigurationRepository(config('explorer.indexes') ?? []);
+            return new ElasticIndexConfigurationRepository(
+                config('explorer.indexes') ?? [],
+                config('explorer.prune_old_aliases')
+            );
         });
 
         resolve(EngineManager::class)->extend('elastic', function (Application $app) {