+++ changelog.md
@@ -7,11 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Added
+- Index aliases, useful for zero downtime indexing
 - The scout:index and scout:delete commands are now implemented for Elasticsearch
 - A DocumentAdapterInterface with an adapter that only deals with documents
 
 ### Changed
 - The IndexAdapterInterface now only focuses on indices
+- The BulkOperation class now requires an instance of IndexConfigurationInterface
 
 ### Deprecated
 - If you rely on the old IndexAdapterInterface, use the DeprecatedElasticAdapterInterface instead of the IndexAdapterInterface to keep the functionality working as it previously was